requirements.txt
asgiref==3.2.3
Django==4.2.5
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0